- move NgDialectConversionPass to anonymous namespace
